@@ -36,12 +36,13 @@
         ]
     },
     "require": {
-        "php": ">=7.1.0",
-        "zendframework/zend-diactoros": "^1.7"
+        "php": "^8.0",
+        "laminas/laminas-diactoros": "^2.24",
+        "rector/rector": "^0.15.19"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.2",
-        "phpstan/phpstan": "^0.10.1",
+        "phpunit/phpunit": "^9.6",
+        "phpstan/phpstan": "^1.0",
         "friendsofphp/php-cs-fixer": "^2.12"
     }
 }
